refactor(constants, utils): split out `getCategoryByTag` and `categoryToTagsMap`
